-added genders to patients and doctors
-Patient & Doctors now have genders.
-On appointment creation, patient can filter doctors by gender.
-If patient is of the opposite gender, a note will be shown that this doctor is of the opposite gender.
-Apponintments now have a new field "been", added this inorder to allow/disallow patients to rate doctors.
If patient did not had any past appointments with this doctor then he cannot review this doctor.
also "been" status which is boolean will only set to "True" if the patient came to the appointment and the doctor updated the status so patients wont create an appointment, the appointment date will pass and then he will be allowed to "Rate" the doctor.